@@ -4,40 +4,69 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Blog.Core.Model;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Blog.Core.Controllers
 {
+    /// <summary>
+    /// 图片管理
+    /// </summary>
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class ImgController : Controller
     {
-        // GET: api/Img
+        // GET: api/Download
+        /// <summary>
+        /// 下载图片（支持中文字符）
+        /// </summary>
+        /// <param name="environment"></param>
+        /// <returns></returns>
         [HttpGet]
-        public IEnumerable<string> Get()
+        [Route("down")]
+        public FileStreamResult DownImg([FromServices]IWebHostEnvironment environment)
         {
-            return new string[] { "value1", "value2" };
+            string foldername = "";
+            string filepath = Path.Combine(environment.WebRootPath, foldername, "测试下载中文名称的图片.png");
+            var stream = System.IO.File.OpenRead(filepath);
+            string fileExt = ".jpg";  // 这里可以写一个获取文件扩展名的方法，获取扩展名
+            //获取文件的ContentType
+            var provider = new Microsoft.AspNetCore.StaticFiles.FileExtensionContentTypeProvider();
+            var memi = provider.Mappings[fileExt];
+            var fileName = Path.GetFileName(filepath);
+
+
+            return File(stream, memi, fileName);
         }
 
-        // GET: api/Img/5
-        [HttpGet("{id}", Name = "Get")]
-        public string Get(int id)
-        {
-            return "value";
-        }
-
-
+        /// <summary>
+        /// 上传图片,多文件，可以使用 postman 测试，
+        /// 如果是单文件，可以 参数写 IFormFile file1
+        /// </summary>
+        /// <param name="environment"></param>
+        /// <returns></returns>
         [HttpPost]
         [Route("Pic")]
-        public async Task<MessageModel<string>> InsertPicture([FromServices]IHostingEnvironment environment)
+        public async Task<MessageModel<string>> InsertPicture([FromServices]IWebHostEnvironment environment)
         {
             var data = new MessageModel<string>();
             string path = string.Empty;
             string foldername = "images";
-            var files = Request.Form.Files;
-            if (files == null || files.Count() <= 0) { data.msg = "请选择上传的文件。"; return data; }
+            IFormFileCollection files = null;
+
+            try
+            {
+                files = Request.Form.Files;
+            }
+            catch (Exception)
+            {
+                files = null;
+            }
+
+            if (files == null || !files.Any()) { data.msg = "请选择上传的文件。"; return data; }
             //格式限制
             var allowType = new string[] { "image/jpg", "image/png", "image/jpeg" };
 
@@ -102,5 +131,5 @@ namespace Blog.Core.Controllers
         {
         }
     }
-  
+
 }
