@@ -7,15 +7,23 @@ using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 using Autofac;
-using Autofac.Extensions.DependencyInjection;
 using Autofac.Extras.DynamicProxy;
 using AutoMapper;
 using Blog.Core.AOP;
 using Blog.Core.AuthHelper;
 using Blog.Core.Common;
+using Blog.Core.Common.AppConfig;
+using Blog.Core.Common.DB;
+using Blog.Core.Common.HttpContextUser;
+using Blog.Core.Common.LogHelper;
+using Blog.Core.Common.MemoryCache;
+using Blog.Core.Extensions;
 using Blog.Core.Filter;
-using Blog.Core.Log;
+using Blog.Core.Hubs;
+using Blog.Core.IServices;
+using Blog.Core.Middlewares;
 using Blog.Core.Model;
+using Blog.Core.Tasks;
 using log4net;
 using log4net.Config;
 using log4net.Repository;
@@ -23,16 +31,18 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
+using Microsoft.OpenApi.Models;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using StackExchange.Profiling.Storage;
-using Swashbuckle.AspNetCore.Swagger;
+using Swashbuckle.AspNetCore.Filters;
 using static Blog.Core.SwaggerHelper.CustomApiVersion;
 
 namespace Blog.Core
@@ -43,36 +53,60 @@ namespace Blog.Core
         /// <summary>
         /// log4net 仓储库
         /// </summary>
-        public static ILoggerRepository repository { get; set; }
-
-        public Startup(IConfiguration configuration)
+        public static ILoggerRepository Repository { get; set; }
+        private static readonly ILog log =
+        LogManager.GetLogger(typeof(GlobalExceptionsFilter));
+        public Startup(IConfiguration configuration, IWebHostEnvironment env)
         {
             Configuration = configuration;
+            Env = env;
             //log4net
-            repository = LogManager.CreateRepository("Blog.Core");
-            //指定配置文件
-            XmlConfigurator.Configure(repository, new FileInfo("log4net.config"));
+            #region 这部分代码迁到 program.cs 文件里了
+            //Repository = LogManager.CreateRepository(Configuration["Logging:Log4Net:Name"]);
+            ////指定配置文件，如果这里你遇到问题，应该是使用了InProcess模式，请查看Blog.Core.csproj,并删之
+            //var contentPath = env.ContentRootPath;
+            //var log4Config = Path.Combine(contentPath, "log4net.config");
+            //XmlConfigurator.Configure(Repository, new FileInfo(log4Config)); 
+            #endregion
 
         }
 
         public IConfiguration Configuration { get; }
+        public IWebHostEnvironment Env { get; }
         private const string ApiName = "Blog.Core";
 
         // This method gets called by the runtime. Use this method to add services to the container.
-        public IServiceProvider ConfigureServices(IServiceCollection services)
+        public void ConfigureServices(IServiceCollection services)
         {
-            #region 部分服务注入-netcore自带方法
-            //缓存注入
+            #region 部分工具类服务注入-netcore原生方法
+            // 缓存注入
             services.AddScoped<ICaching, MemoryCaching>();
             services.AddSingleton<IMemoryCache>(factory =>
             {
                 var cache = new MemoryCache(new MemoryCacheOptions());
                 return cache;
             });
-            //Redis注入
+            // Redis注入
             services.AddSingleton<IRedisCacheManager, RedisCacheManager>();
-            //log日志注入
-            services.AddSingleton<ILoggerHelper, LogHelper>();
+
+            services.AddSingleton(new Appsettings(Env.ContentRootPath));
+
+            services.AddSingleton(new LogLock(Env.ContentRootPath));
+            #endregion
+
+            #region ISqlSugarClient
+
+            services.AddScoped<SqlSugar.ISqlSugarClient>(o =>
+            {
+                return new SqlSugar.SqlSugarClient(new SqlSugar.ConnectionConfig()
+                {
+                    ConnectionString = BaseDBConfig.ConnectionString,//必填, 数据库连接字符串
+                    DbType = (SqlSugar.DbType)BaseDBConfig.DbType,//必填, 数据库类型
+                    IsAutoCloseConnection = true,//默认false, 时候知道关闭数据库连接, 设置为true无需使用using或者Close操作
+                    InitKeyType = SqlSugar.InitKeyType.SystemTable//默认SystemTable, 字段信息读取, 如：该属性是不是主键，标识列等等信息
+                });
+            });
+
             #endregion
 
             #region 初始化DB
@@ -88,23 +122,14 @@ namespace Blog.Core
             //跨域第二种方法，声明策略，记得下边app中配置
             services.AddCors(c =>
             {
-                //↓↓↓↓↓↓↓注意正式环境不要使用这种全开放的处理↓↓↓↓↓↓↓↓↓↓
-                c.AddPolicy("AllRequests", policy =>
-                {
-                    policy
-                    .AllowAnyOrigin()//允许任何源
-                    .AllowAnyMethod()//允许任何方式
-                    .AllowAnyHeader()//允许任何头
-                    .AllowCredentials();//允许cookie
-                });
-                //↑↑↑↑↑↑↑注意正式环境不要使用这种全开放的处理↑↑↑↑↑↑↑↑↑↑
-
 
                 //一般采用这种方法
                 c.AddPolicy("LimitRequests", policy =>
                 {
+                    // 支持多个域名端口，注意端口号后不要带/斜杆：比如localhost:8000/，是错的
+                    // 注意，http://127.0.0.1:1818 和 http://localhost:1818 是不一样的，尽量写两个
                     policy
-                    .WithOrigins("http://127.0.0.1:1818", "http://localhost:8080", "http://localhost:8021", "http://localhost:8081", "http://localhost:1818")//支持多个域名端口，注意端口号后不要带/斜杆：比如localhost:8000/，是错的
+                    .WithOrigins("http://127.0.0.1:1818", "http://localhost:8080", "http://localhost:8021", "http://localhost:8081", "http://localhost:1818")
                     .AllowAnyHeader()//Ensures that the policy allows any header.
                     .AllowAnyMethod();
                 });
@@ -119,8 +144,13 @@ namespace Blog.Core
             services.AddMiniProfiler(options =>
                 {
                     options.RouteBasePath = "/profiler";
-                    (options.Storage as MemoryCacheStorage).CacheDuration = TimeSpan.FromMinutes(10);
+                    //(options.Storage as MemoryCacheStorage).CacheDuration = TimeSpan.FromMinutes(10);
+                    options.PopupRenderPosition = StackExchange.Profiling.RenderPosition.Left;
+                    options.PopupShowTimeWithChildren = true;
 
+                    // 可以增加权限
+                    //options.ResultsAuthorize = request => request.HttpContext.User.IsInRole("Admin");
+                    //options.UserIdProvider = request => request.HttpContext.User.Identity.Name;
                 }
             );
 
@@ -134,44 +164,48 @@ namespace Blog.Core
                 //遍历出全部的版本，做文档信息展示
                 typeof(ApiVersions).GetEnumNames().ToList().ForEach(version =>
                 {
-                    c.SwaggerDoc(version, new Info
+                    c.SwaggerDoc(version, new OpenApiInfo
                     {
                         // {ApiName} 定义成全局变量，方便修改
                         Version = version,
-                        Title = $"{ApiName} 接口文档",
+                        Title = $"{ApiName} 接口文档——Netcore 3.0",
                         Description = $"{ApiName} HTTP API " + version,
-                        TermsOfService = "None",
-                        Contact = new Contact { Name = "Blog.Core", Email = "Blog.Core@xxx.com", Url = "https://www.jianshu.com/u/94102b59cc2a" }
+                        Contact = new OpenApiContact { Name = ApiName, Email = "Blog.Core@xxx.com", Url = new Uri("https://www.jianshu.com/u/94102b59cc2a") },
+                        License = new OpenApiLicense { Name = ApiName, Url = new Uri("https://www.jianshu.com/u/94102b59cc2a") }
                     });
-                    // 按相对路径排序，作者：Alby
                     c.OrderActionsBy(o => o.RelativePath);
                 });
 
 
-                //就是这里
-                var xmlPath = Path.Combine(basePath, "Blog.Core.xml");//这个就是刚刚配置的xml文件名
-                c.IncludeXmlComments(xmlPath, true);//默认的第二个参数是false，这个是controller的注释，记得修改
+                try
+                {
+                    //就是这里
+                    var xmlPath = Path.Combine(basePath, "Blog.Core.xml");//这个就是刚刚配置的xml文件名
+                    c.IncludeXmlComments(xmlPath, true);//默认的第二个参数是false，这个是controller的注释，记得修改
 
-                var xmlModelPath = Path.Combine(basePath, "Blog.Core.Model.xml");//这个就是Model层的xml文件名
-                c.IncludeXmlComments(xmlModelPath);
+                    var xmlModelPath = Path.Combine(basePath, "Blog.Core.Model.xml");//这个就是Model层的xml文件名
+                    c.IncludeXmlComments(xmlModelPath);
+                }
+                catch (Exception ex)
+                {
+                    log.Error("Blog.Core.xml和Blog.Core.Model.xml 丢失，请检查并拷贝。\n" + ex.Message);
+                }
+
+                c.OperationFilter<AddResponseHeadersFilter>();
+                c.OperationFilter<AppendAuthorizeToSummaryOperationFilter>();
+
+                c.OperationFilter<SecurityRequirementsOperationFilter>();
+
 
                 #region Token绑定到ConfigureServices
 
-                //添加header验证信息
-                //c.OperationFilter<SwaggerHeader>();
 
-                // 发行人
-                var IssuerName = (Configuration.GetSection("Audience"))["Issuer"];
-                var security = new Dictionary<string, IEnumerable<string>> { { IssuerName, new string[] { } }, };
-                c.AddSecurityRequirement(security);
-
-                //方案名称“Blog.Core”可自定义，上下一致即可
-                c.AddSecurityDefinition(IssuerName, new ApiKeyScheme
+                c.AddSecurityDefinition("oauth2", new OpenApiSecurityScheme
                 {
                     Description = "JWT授权(数据将在请求头中进行传输) 直接在下框中输入Bearer {token}（注意两者之间是一个空格）\"",
                     Name = "Authorization",//jwt默认的参数名称
-                    In = "header",//jwt默认存放Authorization信息的位置(请求头中)
-                    Type = "apiKey"
+                    In = ParameterLocation.Header,//jwt默认存放Authorization信息的位置(请求头中)
+                    Type = SecuritySchemeType.ApiKey
                 });
                 #endregion
             });
@@ -181,31 +215,96 @@ namespace Blog.Core
             #region MVC + GlobalExceptions
 
             //注入全局异常捕获
-            services.AddMvc(o =>
+            services.AddControllers(o =>
             {
                 // 全局异常过滤
                 o.Filters.Add(typeof(GlobalExceptionsFilter));
+                // 全局路由权限公约
+                //o.Conventions.Insert(0, new GlobalRouteAuthorizeConvention());
+                // 全局路由前缀，统一修改路由
+                o.Conventions.Insert(0, new GlobalRoutePrefixFilter(new RouteAttribute(RoutePrefix.Name)));
             })
-            .SetCompatibilityVersion(CompatibilityVersion.Version_2_2)
-            // 取消默认驼峰
-            .AddJsonOptions(options => { options.SerializerSettings.ContractResolver = new DefaultContractResolver(); });
+            //全局配置Json序列化处理
+            .AddNewtonsoftJson(options =>
+            {
+                //忽略循环引用
+                options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+                //不使用驼峰样式的key
+                options.SerializerSettings.ContractResolver = new DefaultContractResolver();
+                //设置时间格式
+                //options.SerializerSettings.DateFormatString = "yyyy-MM-dd";
+            });
+
+
+
+
 
             #endregion
 
-            #region Authorize权限设置三种情况
+            #region TimedJob
 
-            #region 1、基于角色API授权 + 自定义认证中间件
-
-            // 1【授权】、这个很简单，其他什么都不用做，
-            // 无需配置服务，只需要在API层的controller上边，增加特性即可，注意，只能是角色的:
-            // [Authorize(Roles = "Admin")]
-
-            // 2【认证】、然后在下边的configure里，配置中间件即可:
-            // app.UseMiddleware<JwtTokenAuth>();
+            services.AddHostedService<Job1TimedService>();
+            services.AddHostedService<Job2TimedService>();
 
             #endregion
 
-            #region 2、基于角色的策略授权（简单版） + 自定义认证中间件
+            #region Httpcontext
+
+            // Httpcontext 注入
+            services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
+            services.AddScoped<IUser, AspNetUser>();
+
+            #endregion
+
+            #region SignalR 通讯
+            services.AddSignalR().AddNewtonsoftJsonProtocol();
+            #endregion
+
+            #region Authorize 权限认证三步走
+
+            //Tips：注释中的中括号【xxx】的内容，与下边region中的模块，是一一匹配的
+
+            /*
+             * 如果不想走数据库，仅仅想在代码里配置授权，这里可以按照下边的步骤：
+             * 1步、【1、基于角色的API授权】
+             * 很简单，只需要在指定的接口上，配置特性即可，比如：[Authorize(Roles = "Admin,System,Others")]
+             * 
+             * 但是如果你感觉"Admin,System,Others"，这样的字符串太长的话，可以把这个融合到简单策略里          
+             * 具体的配置，看下文的Region模块【2、基于策略的授权（简单版）】 ，然后在接口上，配置特性：[Authorize(Policy = "A_S_O")]
+             * 
+             * 
+             * 2步、配置Bearer认证服务，具体代码看下文的 region 【第二步：配置认证服务】
+             * 
+             * 3步、开启中间件
+             */
+
+
+
+            /*
+             * 如果想要把权限配置到数据库，步骤如下：
+             * 1步、【3复杂策略授权】
+             * 具体的查看下边 region 内的内容
+             * 
+             * 2步、配置Bearer认证服务，具体代码看下文的 region 【第二步：配置认证服务】
+             * 
+             * 3步、开启中间件
+             */
+
+
+
+            //3、综上所述，设置权限，必须要三步走，授权 + 配置认证服务 + 开启授权中间件，只不过自定义的中间件不能验证过期时间，所以我都是用官方的。
+
+            #region 【第一步：授权】
+
+            #region 1、基于角色的API授权 
+
+            // 1【授权】、这个很简单，其他什么都不用做， 只需要在API层的controller上边，增加特性即可，注意，只能是角色的:
+            // [Authorize(Roles = "Admin,System")]
+
+
+            #endregion
+
+            #region 2、基于策略的授权（简单版）
 
             // 1【授权】、这个和上边的异曲同工，好处就是不用在controller中，写多个 roles 。
             // 然后这么写 [Authorize(Policy = "Admin")]
@@ -214,36 +313,22 @@ namespace Blog.Core
                 options.AddPolicy("Client", policy => policy.RequireRole("Client").Build());
                 options.AddPolicy("Admin", policy => policy.RequireRole("Admin").Build());
                 options.AddPolicy("SystemOrAdmin", policy => policy.RequireRole("Admin", "System"));
+                options.AddPolicy("A_S_O", policy => policy.RequireRole("Admin", "System", "Others"));
             });
 
 
-            // 2【认证】、然后在下边的configure里，配置中间件即可:
-            // app.UseMiddleware<JwtTokenAuth>();
             #endregion
 
-
-            #region 3、复杂策略授权 + 官方JWT认证
+            #region 【3、复杂策略授权】
 
             #region 参数
             //读取配置文件
             var audienceConfig = Configuration.GetSection("Audience");
-            var symmetricKeyAsBase64 = audienceConfig["Secret"];
+            var symmetricKeyAsBase64 = AppSecretConfig.Audience_Secret_String;
             var keyByteArray = Encoding.ASCII.GetBytes(symmetricKeyAsBase64);
             var signingKey = new SymmetricSecurityKey(keyByteArray);
 
-            // 令牌验证参数
-            var tokenValidationParameters = new TokenValidationParameters
-            {
-                ValidateIssuerSigningKey = true,
-                IssuerSigningKey = signingKey,
-                ValidateIssuer = true,
-                ValidIssuer = audienceConfig["Issuer"],//发行人
-                ValidateAudience = true,
-                ValidAudience = audienceConfig["Audience"],//订阅人
-                ValidateLifetime = true,
-                ClockSkew = TimeSpan.Zero,
-                RequireExpirationTime = true,
-            };
+
             var signingCredentials = new SigningCredentials(signingKey, SecurityAlgorithms.HmacSha256);
 
             // 如果要数据库动态绑定，这里先留个空，后边处理器里动态赋值
@@ -257,57 +342,94 @@ namespace Blog.Core
                 audienceConfig["Issuer"],//发行人
                 audienceConfig["Audience"],//听众
                 signingCredentials,//签名凭据
-                expiration: TimeSpan.FromSeconds(60 * 30)//接口的过期时间
-                ); 
+                expiration: TimeSpan.FromSeconds(60 * 60)//接口的过期时间
+                );
             #endregion
 
-            //1【授权】、自定义复杂授权的权限要求
+            //【授权】
             services.AddAuthorization(options =>
             {
-                options.AddPolicy("Permission",
+                options.AddPolicy(Permissions.Name,
                          policy => policy.Requirements.Add(permissionRequirement));
-            })
-            //2【认证】、官方JWT认证
-            .AddAuthentication(x =>
-            {
-                x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            })
-            .AddJwtBearer(o =>
-            {
-                o.TokenValidationParameters = tokenValidationParameters;
-                o.Events = new JwtBearerEvents
-                {
-                    OnAuthenticationFailed = context =>
-                    {
-                        // 如果过期，则把<是否过期>添加到，返回头信息中
-                        if (context.Exception.GetType() == typeof(SecurityTokenExpiredException))
-                        {
-                            context.Response.Headers.Add("Token-Expired", "true");
-                        }
-                        return Task.CompletedTask;
-                    }
-                };
             });
+
+
+            #endregion
+
+
+            #endregion
+
+
+
+
+
+            #region 【第二步：配置认证服务】
+            // 令牌验证参数
+            var tokenValidationParameters = new TokenValidationParameters
+            {
+                ValidateIssuerSigningKey = true,
+                IssuerSigningKey = signingKey,
+                ValidateIssuer = true,
+                ValidIssuer = audienceConfig["Issuer"],//发行人
+                ValidateAudience = true,
+                ValidAudience = audienceConfig["Audience"],//订阅人
+                ValidateLifetime = true,
+                ClockSkew = TimeSpan.FromSeconds(30),
+                RequireExpirationTime = true,
+            };
+
+            //2.1【认证】、core自带官方JWT认证
+            // 开启Bearer认证
+            services.AddAuthentication("Bearer")
+             // 添加JwtBearer服务
+             .AddJwtBearer(o =>
+             {
+                 o.TokenValidationParameters = tokenValidationParameters;
+                 o.Events = new JwtBearerEvents
+                 {
+                     OnAuthenticationFailed = context =>
+                     {
+                         // 如果过期，则把<是否过期>添加到，返回头信息中
+                         if (context.Exception.GetType() == typeof(SecurityTokenExpiredException))
+                         {
+                             context.Response.Headers.Add("Token-Expired", "true");
+                         }
+                         return Task.CompletedTask;
+                     }
+                 };
+             });
+
+
+            //2.2【认证】、IdentityServer4 认证 (暂时忽略)
+            //services.AddAuthentication("Bearer")
+            //  .AddIdentityServerAuthentication(options =>
+            //  {
+            //      options.Authority = "http://localhost:5002";
+            //      options.RequireHttpsMetadata = false;
+            //      options.ApiName = "blog.core.api";
+            //  });
+            // 注入权限处理器
 
             services.AddSingleton<IAuthorizationHandler, PermissionHandler>();
             services.AddSingleton(permissionRequirement);
+            #endregion
 
             #endregion
 
+        }
 
+        public void ConfigureContainer(ContainerBuilder builder)
+        {
+            var basePath = Microsoft.DotNet.PlatformAbstractions.ApplicationEnvironment.ApplicationBasePath;
 
-            #endregion
-
-
-            #region AutoFac DI
-            //实例化 AutoFac  容器   
-            var builder = new ContainerBuilder();
             //注册要通过反射创建的组件
             //builder.RegisterType<AdvertisementServices>().As<IAdvertisementServices>();
             builder.RegisterType<BlogCacheAOP>();//可以直接替换其他拦截器
             builder.RegisterType<BlogRedisCacheAOP>();//可以直接替换其他拦截器
             builder.RegisterType<BlogLogAOP>();//这样可以注入第二个
+            builder.RegisterType<BlogRedisCacheAOP>();
+            builder.RegisterType<BlogLogAOP>();
+            builder.RegisterType<BlogTranAOP>();
 
             // ※※★※※ 如果你是第一次下载项目，请先F6编译，然后再F5执行，※※★※※
 
@@ -315,23 +437,53 @@ namespace Blog.Core
 
             #region Service.dll 注入，有对应接口
             //获取项目绝对路径，请注意，这个是实现类的dll文件，不是接口 IService.dll ，注入容器当然是Activatore
-            var servicesDllFile = Path.Combine(basePath, "Blog.Core.Services.dll");
-            var assemblysServices = Assembly.LoadFile(servicesDllFile);//直接采用加载文件的方法
+            try
+            {
+                var servicesDllFile = Path.Combine(basePath, "Blog.Core.Services.dll");
+                var assemblysServices = Assembly.LoadFrom(servicesDllFile);//直接采用加载文件的方法  ※※★※※ 如果你是第一次下载项目，请先F6编译，然后再F5执行，※※★※※
 
-            //builder.RegisterAssemblyTypes(assemblysServices).AsImplementedInterfaces();//指定已扫描程序集中的类型注册为提供所有其实现的接口。
+                //builder.RegisterAssemblyTypes(assemblysServices).AsImplementedInterfaces();//指定已扫描程序集中的类型注册为提供所有其实现的接口。
 
-            builder.RegisterAssemblyTypes(assemblysServices)
-                      .AsImplementedInterfaces()
-                      .InstancePerLifetimeScope()
-                      .EnableInterfaceInterceptors()//引用Autofac.Extras.DynamicProxy;
-                                                    // 如果你想注入两个，就这么写  InterceptedBy(typeof(BlogCacheAOP), typeof(BlogLogAOP));
-                      .InterceptedBy(typeof(BlogRedisCacheAOP));//允许将拦截器服务的列表分配给注册。 
-            #endregion
+
+                // AOP 开关，如果想要打开指定的功能，只需要在 appsettigns.json 对应对应 true 就行。
+                var cacheType = new List<Type>();
+                if (Appsettings.app(new string[] { "AppSettings", "RedisCachingAOP", "Enabled" }).ObjToBool())
+                {
+                    cacheType.Add(typeof(BlogRedisCacheAOP));
+                }
+                if (Appsettings.app(new string[] { "AppSettings", "MemoryCachingAOP", "Enabled" }).ObjToBool())
+                {
+                    cacheType.Add(typeof(BlogCacheAOP));
+                }
+                if (Appsettings.app(new string[] { "AppSettings", "TranAOP", "Enabled" }).ObjToBool())
+                {
+                    cacheType.Add(typeof(BlogTranAOP));
+                }
+                if (Appsettings.app(new string[] { "AppSettings", "LogAOP", "Enabled" }).ObjToBool())
+                {
+                    cacheType.Add(typeof(BlogLogAOP));
+                }
+
+                builder.RegisterAssemblyTypes(assemblysServices)
+                          .AsImplementedInterfaces()
+                          .InstancePerLifetimeScope()
+                          .EnableInterfaceInterceptors()//引用Autofac.Extras.DynamicProxy;
+                                                        // 如果你想注入两个，就这么写  InterceptedBy(typeof(BlogCacheAOP), typeof(BlogLogAOP));
+                                                        // 如果想使用Redis缓存，请必须开启 redis 服务，端口号我的是6319，如果不一样还是无效，否则请使用memory缓存 BlogCacheAOP
+                          .InterceptedBy(cacheType.ToArray());//允许将拦截器服务的列表分配给注册。 
+                #endregion
 
             #region Repository.dll 注入，有对应接口
-            var repositoryDllFile = Path.Combine(basePath, "Blog.Core.Repository.dll");
-            var assemblysRepository = Assembly.LoadFile(repositoryDllFile);
-            builder.RegisterAssemblyTypes(assemblysRepository).AsImplementedInterfaces();
+                var repositoryDllFile = Path.Combine(basePath, "Blog.Core.Repository.dll");
+                var assemblysRepository = Assembly.LoadFrom(repositoryDllFile);
+                builder.RegisterAssemblyTypes(assemblysRepository).AsImplementedInterfaces();
+            }
+            catch (Exception ex)
+            {
+                log.Error("Repository.dll和service.dll 丢失，因为项目解耦了，所以需要先F6编译，再F5运行，请检查并拷贝。\n" + ex.Message);
+
+                //throw new Exception("※※★※※ 如果你是第一次下载项目，请先对整个解决方案dotnet build（F6编译），然后再对api层 dotnet run（F5执行），\n因为解耦了，如果你是发布的模式，请检查bin文件夹是否存在Repository.dll和service.dll ※※★※※" + ex.Message + "\n" + ex.InnerException);
+            }
             #endregion
             #endregion
 
@@ -339,6 +491,7 @@ namespace Blog.Core
             #region 没有接口层的服务层注入
 
             ////因为没有接口层，所以不能实现解耦，只能用 Load 方法。
+            ////注意如果使用没有接口的服务，并想对其使用 AOP 拦截，就必须设置为虚方法
             ////var assemblysServicesNoInterfaces = Assembly.Load("Blog.Core.Services");
             ////builder.RegisterAssemblyTypes(assemblysServicesNoInterfaces);
 
@@ -352,27 +505,47 @@ namespace Blog.Core
 
             #endregion
 
-
-            //将services填充到Autofac容器生成器中
-            builder.Populate(services);
-
             //使用已进行的组件登记创建新容器
-            var ApplicationContainer = builder.Build();
-
-            #endregion
-
-            return new AutofacServiceProvider(ApplicationContainer);//第三方IOC接管 core内置DI容器
+            //var ApplicationContainer = builder.Build();
 
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IBlogArticleServices _blogArticleServices, ILoggerFactory loggerFactory)
         {
+
+            #region RecordAllLogs
+
+            if (Appsettings.app("AppSettings", "Middleware_RecordAllLogs", "Enabled").ObjToBool())
+            {
+                loggerFactory.AddLog4Net();//记录所有的访问记录
+            }
+
+            #endregion
+
+            #region ReuestResponseLog
+
+            if (Appsettings.app("AppSettings", "Middleware_RequestResponse", "Enabled").ObjToBool())
+            {
+                app.UseReuestResponseLog();//记录请求与返回数据 
+            }
+
+            #endregion
+
             #region Environment
             if (env.IsDevelopment())
             {
                 // 在开发环境中，使用异常页面，这样可以暴露错误堆栈信息，所以不要放在生产环境。
                 app.UseDeveloperExceptionPage();
+
+                //app.Use(async (context, next) =>
+                //{
+                //    //这里会多次调用，这里测试一下就行，不要打开注释
+                //    //var blogs =await _blogArticleServices.GetBlogs();
+                //    var processName = System.Diagnostics.Process.GetCurrentProcess().ProcessName;
+                //    Console.WriteLine(processName);
+                //    await next();
+                //});
             }
             else
             {
@@ -384,10 +557,6 @@ namespace Blog.Core
             }
             #endregion
 
-            #region MiniProfiler
-            app.UseMiniProfiler();
-            #endregion
-
             #region Swagger
             app.UseSwagger();
             app.UseSwaggerUI(c =>
@@ -397,17 +566,14 @@ namespace Blog.Core
                 {
                     c.SwaggerEndpoint($"/swagger/{version}/swagger.json", $"{ApiName} {version}");
                 });
-                c.IndexStream = () => GetType().GetTypeInfo().Assembly.GetManifestResourceStream("Blog.Core.index.html");
-                c.RoutePrefix = ""; //路径配置，设置为空，表示直接在根域名（localhost:8001）访问该文件,注意localhost:8001/swagger是访问不到的，去launchSettings.json把launchUrl去掉
+                // 将swagger首页，设置成我们自定义的页面，记得这个字符串的写法：解决方案名.index.html
+                c.IndexStream = () => GetType().GetTypeInfo().Assembly.GetManifestResourceStream("Blog.Core.index.html");//这里是配合MiniProfiler进行性能监控的，《文章：完美基于AOP的接口性能分析》，如果你不需要，可以暂时先注释掉，不影响大局。
+                c.RoutePrefix = ""; //路径配置，设置为空，表示直接在根域名（localhost:8001）访问该文件,注意localhost:8001/swagger是访问不到的，去launchSettings.json把launchUrl去掉，如果你想换一个路径，直接写名字即可，比如直接写c.RoutePrefix = "doc";
             });
             #endregion
 
-            #region Authen
-
-            //app.UseMiddleware<JwtTokenAuth>();//此授权认证方法已经放弃，请使用下边的官方验证方法。但是如果你还想传User的全局变量，还是可以继续使用中间件
-
-            // 如果你想使用官方认证，必须在上边ConfigureService 中，配置JWT的认证服务 (.AddAuthentication 和 .AddJwtBearer 二者缺一不可)
-            app.UseAuthentication();
+            #region MiniProfiler
+            app.UseMiniProfiler();
             #endregion
 
             #region CORS
@@ -424,7 +590,7 @@ namespace Blog.Core
             #endregion
 
             // 跳转https
-            app.UseHttpsRedirection();
+            //app.UseHttpsRedirection();
             // 使用静态文件
             app.UseStaticFiles();
             // 使用cookie
@@ -432,7 +598,31 @@ namespace Blog.Core
             // 返回错误码
             app.UseStatusCodePages();//把错误码返回前台，比如是404
 
-            app.UseMvc();
+
+
+            app.UseRouting();
+
+            #region 第三步：开启认证中间件
+
+            //此授权认证方法已经放弃，请使用下边的官方验证方法。但是如果你还想传User的全局变量，还是可以继续使用中间件，第二种写法//app.UseMiddleware<JwtTokenAuth>(); 
+            //app.UseJwtTokenAuth();
+
+            //如果你想使用官方认证，必须在上边ConfigureService 中，配置JWT的认证服务 (.AddAuthentication 和 .AddJwtBearer 二者缺一不可)
+            app.UseAuthentication();
+            #endregion
+
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute(
+                    name: "default",
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
+
+                endpoints.MapHub<ChatHub>("/api2/chatHub");
+            });
+
+
         }
 
     }
