@@ -1,7 +1,10 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Configuration.Json;
 using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using System.Text;
 
 namespace Blog.Core.Common
@@ -12,35 +15,41 @@ namespace Blog.Core.Common
     public class Appsettings
     {
         static IConfiguration Configuration { get; set; }
-        static Appsettings()
+        static string contentPath { get; set; }
+
+        public Appsettings(string contentPath)
         {
-            //ReloadOnChange = true 当appsettings.json被修改时重新加载
+            string Path = "appsettings.json";
+
+            //如果你把配置文件 是 根据环境变量来分开了，可以这样写
+            //Path = $"appsettings.{Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT")}.json";
+
             Configuration = new ConfigurationBuilder()
-            .Add(new JsonConfigurationSource { Path = "appsettings.json", ReloadOnChange = true })
-            .Build();
+               .SetBasePath(contentPath)
+               .Add(new JsonConfigurationSource { Path = Path, Optional = false, ReloadOnChange = true })//这样的话，可以直接读目录里的json文件，而不是 bin 文件夹下的，所以不用修改复制属性
+               .Build();
+
+
         }
+
         /// <summary>
         /// 封装要操作的字符
         /// </summary>
-        /// <param name="sections"></param>
+        /// <param name="sections">节点配置</param>
         /// <returns></returns>
         public static string app(params string[] sections)
         {
             try
             {
-                var val = string.Empty;
-                for (int i = 0; i < sections.Length; i++)
+
+                if (sections.Any())
                 {
-                    val += sections[i] + ":";
+                    return Configuration[string.Join(":", sections)];
                 }
-
-                return Configuration[val.TrimEnd(':')];
             }
-            catch (Exception)
-            {
-                return "";
-            }
+            catch (Exception) { }
 
+            return "";
         }
     }
 }
