@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Blog.Core.AOP
+﻿namespace Blog.Core.Common.MemoryCache
 {
     /// <summary>
     /// 简单的缓存接口，只有查询和添加，以后会进行扩展
