@@ -1,10 +1,7 @@
 ﻿using Microsoft.Extensions.Caching.Memory;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace Blog.Core.AOP
+namespace Blog.Core.Common.MemoryCache
 {
     /// <summary>
     /// 实例化缓存接口ICaching
@@ -12,7 +9,7 @@ namespace Blog.Core.AOP
     public class MemoryCaching : ICaching
     {
         //引用Microsoft.Extensions.Caching.Memory;这个和.net 还是不一样，没有了Httpruntime了
-        private IMemoryCache _cache;
+        private readonly IMemoryCache _cache;
         //还是通过构造函数的方法，获取
         public MemoryCaching(IMemoryCache cache)
         {
