@@ -14,7 +14,7 @@ namespace Blog.Core.Common
 
         public RedisCacheManager()
         {
-            string redisConfiguration = Appsettings.app(new string[] { "AppSettings", "RedisCaching", "ConnectionString" });//获取连接字符串
+            string redisConfiguration = Appsettings.app(new string[] { "AppSettings", "RedisCachingAOP", "ConnectionString" });//获取连接字符串
 
             if (string.IsNullOrWhiteSpace(redisConfiguration))
             {
@@ -50,8 +50,7 @@ namespace Blog.Core.Common
                 }
                 catch (Exception)
                 {
-
-                    throw new Exception("Redis服务未启用，请开启该服务");
+                    //throw new Exception("Redis服务未启用，请开启该服务，并且请注意端口号，本项目使用的的6319，而且我的是没有设置密码。");
                 }
             }
             return this.redisConnection;
