@@ -10,7 +10,7 @@ namespace Blog.Core.Model
     public class PageModel<T>
     {
         /// <summary>
-        /// 当前页数
+        /// 当前页标
         /// </summary>
         public int page { get; set; } = 1;
         /// <summary>
@@ -22,9 +22,14 @@ namespace Blog.Core.Model
         /// </summary>
         public int dataCount { get; set; } = 0;
         /// <summary>
+        /// 每页大小
+        /// </summary>
+        public int PageSize { set; get; }
+        /// <summary>
         /// 返回数据
         /// </summary>
         public List<T> data { get; set; }
 
     }
+
 }
