@@ -15,23 +15,23 @@ namespace Blog.Core.Services
 {
     public class BlogArticleServices : BaseServices<BlogArticle>, IBlogArticleServices
     {
-        IBlogArticleRepository dal;
-        IMapper IMapper;
-        public BlogArticleServices(IBlogArticleRepository dal, IMapper IMapper)
+        IBlogArticleRepository _dal;
+        IMapper _mapper;
+        public BlogArticleServices(IBlogArticleRepository dal, IMapper mapper)
         {
-            this.dal = dal;
-            base.baseDal = dal;
-            this.IMapper = IMapper;
+            this._dal = dal;
+            base.BaseDal = dal;
+            this._mapper = mapper;
         }
         /// <summary>
         /// 获取视图博客详情信息
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<BlogViewModels> getBlogDetails(int id)
+        public async Task<BlogViewModels> GetBlogDetails(int id)
         {
-            var bloglist = await dal.Query(a => a.bID > 0, a => a.bID);
-            var blogArticle = (await dal.Query(a => a.bID == id)).FirstOrDefault();
+            var bloglist = await base.Query(a => a.bID > 0, a => a.bID);
+            var blogArticle = (await base.Query(a => a.bID == id)).FirstOrDefault();
 
             BlogViewModels models = null;
 
@@ -51,7 +51,7 @@ namespace Blog.Core.Services
 
 
                         // 注意就是这里,mapper
-                        models = IMapper.Map<BlogViewModels>(blogArticle);
+                        models = _mapper.Map<BlogViewModels>(blogArticle);
 
                         if (nextblog != null)
                         {
@@ -64,14 +64,15 @@ namespace Blog.Core.Services
                             models.previous = prevblog.btitle;
                             models.previousID = prevblog.bID;
                         }
+                        var entity2Viewmodel = _mapper.Map<BlogArticle>(models);
 
                     }
-                    catch (Exception) { }
+                    catch (Exception ex) { throw new Exception(ex.Message); }
                 }
 
 
                 blogArticle.btraffic += 1;
-                await dal.Update(blogArticle, new List<string> { "btraffic" });
+                //await base.Update(blogArticle, new List<string> { "btraffic" });
             }
 
             return models;
@@ -85,9 +86,9 @@ namespace Blog.Core.Services
         /// <param name="id"></param>
         /// <returns></returns>
         [Caching(AbsoluteExpiration = 10)]
-        public async Task<List<BlogArticle>> getBlogs()
+        public async Task<List<BlogArticle>> GetBlogs()
         {
-            var bloglist = await dal.Query(a => a.bID > 0, a => a.bID);
+            var bloglist = await base.Query(a => a.bID > 0, a => a.bID);
 
             return bloglist;
 
