@@ -5,6 +5,7 @@ using Blog.Core.IRepository;
 using System.Threading.Tasks;
 using System.Collections.Generic;
 using Blog.Core.Common;
+using System.Linq;
 
 namespace Blog.Core.Services
 {
@@ -13,18 +14,17 @@ namespace Blog.Core.Services
     /// </summary>	
     public class RoleModulePermissionServices : BaseServices<RoleModulePermission>, IRoleModulePermissionServices
     {
-
-        IRoleModulePermissionRepository dal;
-        IModuleRepository moduleRepository;
-        IRoleRepository roleRepository;
+        readonly IRoleModulePermissionRepository _dal;
+        readonly IModuleRepository _moduleRepository;
+        readonly IRoleRepository _roleRepository;
 
         // 将多个仓储接口注入
         public RoleModulePermissionServices(IRoleModulePermissionRepository dal, IModuleRepository moduleRepository, IRoleRepository roleRepository)
         {
-            this.dal = dal;
-            this.moduleRepository = moduleRepository;
-            this.roleRepository = roleRepository;
-            base.baseDal = dal;
+            this._dal = dal;
+            this._moduleRepository = moduleRepository;
+            this._roleRepository = roleRepository;
+            base.BaseDal = dal;
         }
 
         /// <summary>
@@ -34,13 +34,25 @@ namespace Blog.Core.Services
         [Caching(AbsoluteExpiration = 10)]
         public async Task<List<RoleModulePermission>> GetRoleModule()
         {
-            var roleModulePermissions = await dal.Query(a => a.IsDeleted == false);
+            var roleModulePermissions = await base.Query(a => a.IsDeleted == false);
+            var roles = await _roleRepository.Query(a => a.IsDeleted == false);
+            var modules = await _moduleRepository.Query(a => a.IsDeleted == false);
+
+            //var roleModulePermissionsAsync = base.Query(a => a.IsDeleted == false);
+            //var rolesAsync = _roleRepository.Query(a => a.IsDeleted == false);
+            //var modulesAsync = _moduleRepository.Query(a => a.IsDeleted == false);
+
+            //var roleModulePermissions = await roleModulePermissionsAsync;
+            //var roles = await rolesAsync;
+            //var modules = await modulesAsync;
+
+
             if (roleModulePermissions.Count > 0)
             {
                 foreach (var item in roleModulePermissions)
                 {
-                    item.Role = await roleRepository.QueryByID(item.RoleId);
-                    item.Module = await moduleRepository.QueryByID(item.ModuleId);
+                    item.Role = roles.FirstOrDefault(d => d.Id == item.RoleId);
+                    item.Module = modules.FirstOrDefault(d => d.Id == item.ModuleId);
                 }
 
             }
@@ -49,7 +61,11 @@ namespace Blog.Core.Services
 
         public async Task<List<RoleModulePermission>> TestModelWithChildren()
         {
-            return await dal.WithChildrenModel();
+            return await _dal.WithChildrenModel();
+        }
+        public async Task<List<TestMuchTableResult>> QueryMuchTable()
+        {
+            return await _dal.QueryMuchTable();
         }
     }
 }
