@@ -14,11 +14,11 @@ namespace Blog.Core.Services
     public class TopicServices: BaseServices<Topic>, ITopicServices
     {
 
-        ITopicRepository dal;
+        ITopicRepository _dal;
         public TopicServices(ITopicRepository dal)
         {
-            this.dal = dal;
-            base.baseDal = dal;
+            this._dal = dal;
+            base.BaseDal = dal;
         }
 
         /// <summary>
@@ -28,7 +28,7 @@ namespace Blog.Core.Services
         [Caching(AbsoluteExpiration = 60)]
         public async Task<List<Topic>> GetTopics()
         {
-            return await dal.Query(a => !a.tIsDelete && a.tSectendDetail == "tbug");
+            return await base.Query(a => !a.tIsDelete && a.tSectendDetail == "tbug");
         }
 
     }
