@@ -14,11 +14,11 @@ namespace Blog.Core.Services
 	public class UserRoleServices : BaseServices<UserRole>, IUserRoleServices
     {
 	
-        IUserRoleRepository dal;
+        IUserRoleRepository _dal;
         public UserRoleServices(IUserRoleRepository dal)
         {
-            this.dal = dal;
-            base.baseDal = dal;
+            this._dal = dal;
+            base.BaseDal = dal;
         }
         /// <summary>
         /// 
@@ -31,15 +31,15 @@ namespace Blog.Core.Services
             UserRole userRole = new UserRole(uid, rid);
 
             UserRole model = new UserRole();
-            var userList = await dal.Query(a => a.UserId == userRole.UserId && a.RoleId == userRole.RoleId);
+            var userList = await base.Query(a => a.UserId == userRole.UserId && a.RoleId == userRole.RoleId);
             if (userList.Count > 0)
             {
                 model = userList.FirstOrDefault();
             }
             else
             {
-                var id = await dal.Add(userRole);
-                model = await dal.QueryByID(id);
+                var id = await base.Add(userRole);
+                model = await base.QueryById(id);
             }
 
             return model;
@@ -51,7 +51,7 @@ namespace Blog.Core.Services
         [Caching(AbsoluteExpiration = 30)]
         public async Task<int> GetRoleIdByUid(int uid)
         {
-            return ((await dal.Query(d => d.UserId == uid)).OrderByDescending(d => d.Id).LastOrDefault()?.RoleId).ObjToInt();
+            return ((await base.Query(d => d.UserId == uid)).OrderByDescending(d => d.Id).LastOrDefault()?.RoleId).ObjToInt();
         }
     }
 }
