@@ -10,38 +10,38 @@ namespace Blog.Core.FrameWork.Services
     /// <summary>
     /// sysUserInfoServices
     /// </summary>	
-    public class sysUserInfoServices : BaseServices<sysUserInfo>, IsysUserInfoServices
+    public class SysUserInfoServices : BaseServices<sysUserInfo>, ISysUserInfoServices
     {
 
-        IsysUserInfoRepository dal;
-        IUserRoleServices userRoleServices;
-        IRoleRepository roleRepository;
-        public sysUserInfoServices(IsysUserInfoRepository dal, IUserRoleServices userRoleServices, IRoleRepository roleRepository)
+        IsysUserInfoRepository _dal;
+        IUserRoleServices _userRoleServices;
+        IRoleRepository _roleRepository;
+        public SysUserInfoServices(IsysUserInfoRepository dal, IUserRoleServices userRoleServices, IRoleRepository roleRepository)
         {
-            this.dal = dal;
-            this.userRoleServices = userRoleServices;
-            this.roleRepository = roleRepository;
-            base.baseDal = dal;
+            this._dal = dal;
+            this._userRoleServices = userRoleServices;
+            this._roleRepository = roleRepository;
+            base.BaseDal = dal;
         }
         /// <summary>
         /// 
         /// </summary>
         /// <param name="loginName"></param>
-        /// <param name="loginPWD"></param>
+        /// <param name="loginPwd"></param>
         /// <returns></returns>
-        public async Task<sysUserInfo> SaveUserInfo(string loginName, string loginPWD)
+        public async Task<sysUserInfo> SaveUserInfo(string loginName, string loginPwd)
         {
-            sysUserInfo sysUserInfo = new sysUserInfo(loginName, loginPWD);
+            sysUserInfo sysUserInfo = new sysUserInfo(loginName, loginPwd);
             sysUserInfo model = new sysUserInfo();
-            var userList = await dal.Query(a => a.uLoginName == sysUserInfo.uLoginName && a.uLoginPWD == sysUserInfo.uLoginPWD);
+            var userList = await base.Query(a => a.uLoginName == sysUserInfo.uLoginName && a.uLoginPWD == sysUserInfo.uLoginPWD);
             if (userList.Count > 0)
             {
                 model = userList.FirstOrDefault();
             }
             else
             {
-                var id = await dal.Add(sysUserInfo);
-                model = await dal.QueryByID(id);
+                var id = await base.Add(sysUserInfo);
+                model = await base.QueryById(id);
             }
 
             return model;
@@ -52,18 +52,20 @@ namespace Blog.Core.FrameWork.Services
         /// 
         /// </summary>
         /// <param name="loginName"></param>
-        /// <param name="loginPWD"></param>
+        /// <param name="loginPwd"></param>
         /// <returns></returns>
-        public async Task<string> GetUserRoleNameStr(string loginName, string loginPWD)
+        public async Task<string> GetUserRoleNameStr(string loginName, string loginPwd)
         {
             string roleName = "";
-            var user = (await dal.Query(a => a.uLoginName == loginName && a.uLoginPWD == loginPWD)).FirstOrDefault();
+            var user = (await base.Query(a => a.uLoginName == loginName && a.uLoginPWD == loginPwd)).FirstOrDefault();
+            var roleList = await _roleRepository.Query(a => a.IsDeleted==false);
             if (user != null)
             {
-                var userRoles = await userRoleServices.Query(ur => ur.UserId == user.uID);
+                var userRoles = await _userRoleServices.Query(ur => ur.UserId == user.uID);
                 if (userRoles.Count > 0)
                 {
-                    var roles = await roleRepository.QueryByIDs(userRoles.Select(ur => ur.RoleId.ObjToString()).ToArray());
+                    var arr = userRoles.Select(ur => ur.RoleId.ObjToString()).ToList();
+                    var roles = roleList.Where(d => arr.Contains(d.Id.ObjToString()));
 
                     roleName = string.Join(',', roles.Select(r => r.Name).ToArray());
                 }
