@@ -1,40 +1,46 @@
-﻿using Blog.Core.Common;
-using Microsoft.IdentityModel.Tokens;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Security.Claims;
 using System.Text;
-using System.Threading.Tasks;
+using Blog.Core.Common;
+using Blog.Core.Common.AppConfig;
+using Microsoft.IdentityModel.Tokens;
 
-namespace Blog.Core.AuthHelper
+namespace Blog.Core.AuthHelper.OverWrite
 {
     public class JwtHelper
     {
-       
+
         /// <summary>
         /// 颁发JWT字符串
         /// </summary>
         /// <param name="tokenModel"></param>
         /// <returns></returns>
-        public static string IssueJWT(TokenModelJWT tokenModel)
+        public static string IssueJwt(TokenModelJwt tokenModel)
         {
-            var dateTime = DateTime.UtcNow;
-
             string iss = Appsettings.app(new string[] { "Audience", "Issuer" });
             string aud = Appsettings.app(new string[] { "Audience", "Audience" });
-            string secret = Appsettings.app(new string[] { "Audience", "Secret" });
+            string secret = AppSecretConfig.Audience_Secret_String;
 
             //var claims = new Claim[] //old
             var claims = new List<Claim>
                 {
-                    //下边为Claim的默认配置
+                 /*
+                 * 特别重要：
+                   1、这里将用户的部分信息，比如 uid 存到了Claim 中，如果你想知道如何在其他地方将这个 uid从 Token 中取出来，请看下边的SerializeJwt() 方法，或者在整个解决方案，搜索这个方法，看哪里使用了！
+                   2、你也可以研究下 HttpContext.User.Claims ，具体的你可以看看 Policys/PermissionHandler.cs 类中是如何使用的。
+                 */
+
+                    
+
                 new Claim(JwtRegisteredClaimNames.Jti, tokenModel.Uid.ToString()),
                 new Claim(JwtRegisteredClaimNames.Iat, $"{new DateTimeOffset(DateTime.Now).ToUnixTimeSeconds()}"),
                 new Claim(JwtRegisteredClaimNames.Nbf,$"{new DateTimeOffset(DateTime.Now).ToUnixTimeSeconds()}") ,
-                //这个就是过期时间，目前是过期100秒，可自定义，注意JWT有自己的缓冲过期时间
-                new Claim (JwtRegisteredClaimNames.Exp,$"{new DateTimeOffset(DateTime.Now.AddSeconds(100)).ToUnixTimeSeconds()}"),
+                //这个就是过期时间，目前是过期1000秒，可自定义，注意JWT有自己的缓冲过期时间
+                new Claim (JwtRegisteredClaimNames.Exp,$"{new DateTimeOffset(DateTime.Now.AddSeconds(1000)).ToUnixTimeSeconds()}"),
+                new Claim(ClaimTypes.Expiration, DateTime.Now.AddSeconds(1000).ToString()),
                 new Claim(JwtRegisteredClaimNames.Iss,iss),
                 new Claim(JwtRegisteredClaimNames.Aud,aud),
                 
@@ -47,7 +53,7 @@ namespace Blog.Core.AuthHelper
 
 
 
-            //秘钥
+            //秘钥 (SymmetricSecurityKey 对安全性的要求，密钥的长度太短会报出异常)
             var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(secret));
             var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
 
@@ -67,11 +73,11 @@ namespace Blog.Core.AuthHelper
         /// </summary>
         /// <param name="jwtStr"></param>
         /// <returns></returns>
-        public static TokenModelJWT SerializeJWT(string jwtStr)
+        public static TokenModelJwt SerializeJwt(string jwtStr)
         {
             var jwtHandler = new JwtSecurityTokenHandler();
             JwtSecurityToken jwtToken = jwtHandler.ReadJwtToken(jwtStr);
-            object role = new object(); ;
+            object role;
             try
             {
                 jwtToken.Payload.TryGetValue(ClaimTypes.Role, out role);
@@ -81,7 +87,7 @@ namespace Blog.Core.AuthHelper
                 Console.WriteLine(e);
                 throw;
             }
-            var tm = new TokenModelJWT
+            var tm = new TokenModelJwt
             {
                 Uid = (jwtToken.Id).ObjToInt(),
                 Role = role != null ? role.ObjToString() : "",
@@ -93,7 +99,7 @@ namespace Blog.Core.AuthHelper
     /// <summary>
     /// 令牌
     /// </summary>
-    public class TokenModelJWT
+    public class TokenModelJwt
     {
         /// <summary>
         /// Id
