@@ -1,10 +1,6 @@
 ﻿using AutoMapper;
 using Blog.Core.Model.Models;
 using Blog.Core.Model.ViewModels;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Blog.Core.AutoMapper
 {
@@ -16,6 +12,7 @@ namespace Blog.Core.AutoMapper
         public CustomProfile()
         {
             CreateMap<BlogArticle, BlogViewModels>();
+            CreateMap<BlogViewModels, BlogArticle>();
         }
     }
 }
