@@ -5,8 +5,11 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Blog.Core.AuthHelper;
+using Blog.Core.AuthHelper.OverWrite;
+using Blog.Core.Common.Helper;
 using Blog.Core.IServices;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -14,17 +17,17 @@ using Microsoft.AspNetCore.Mvc;
 namespace Blog.Core.Controllers
 {
     /// <summary>
-    /// 
+    /// 登录管理【无权限】
     /// </summary>
     [Produces("application/json")]
     [Route("api/Login")]
+    [AllowAnonymous]
     public class LoginController : Controller
     {
-
-        IsysUserInfoServices sysUserInfoServices;
-        IUserRoleServices userRoleServices;
-        IRoleServices roleServices;
-        PermissionRequirement _requirement;
+        readonly ISysUserInfoServices _sysUserInfoServices;
+        readonly IUserRoleServices _userRoleServices;
+        readonly IRoleServices _roleServices;
+        readonly PermissionRequirement _requirement;
 
 
         /// <summary>
@@ -34,40 +37,37 @@ namespace Blog.Core.Controllers
         /// <param name="userRoleServices"></param>
         /// <param name="roleServices"></param>
         /// <param name="requirement"></param>
-        public LoginController(IsysUserInfoServices sysUserInfoServices, IUserRoleServices userRoleServices, IRoleServices roleServices, PermissionRequirement requirement)
+        public LoginController(ISysUserInfoServices sysUserInfoServices, IUserRoleServices userRoleServices, IRoleServices roleServices, PermissionRequirement requirement)
         {
-            this.sysUserInfoServices = sysUserInfoServices;
-            this.userRoleServices = userRoleServices;
-            this.roleServices = roleServices;
+            this._sysUserInfoServices = sysUserInfoServices;
+            this._userRoleServices = userRoleServices;
+            this._roleServices = roleServices;
             _requirement = requirement;
         }
 
 
         #region 获取token的第1种方法
         /// <summary>
-        /// 获取JWT的方法
+        /// 获取JWT的方法1
         /// </summary>
-        /// <param name="id">id</param>
-        /// <param name="sub">角色</param>
+        /// <param name="name"></param>
+        /// <param name="pass"></param>
         /// <returns></returns>
         [HttpGet]
         [Route("Token")]
-        public async Task<object> GetJWTStr(string name, string pass)
+        public async Task<object> GetJwtStr(string name, string pass)
         {
             string jwtStr = string.Empty;
             bool suc = false;
             //这里就是用户登陆以后，通过数据库去调取数据，分配权限的操作
-            //这里直接写死了
 
-            var user = await sysUserInfoServices.GetUserRoleNameStr(name, pass);
+            var user = await _sysUserInfoServices.GetUserRoleNameStr(name, MD5Helper.MD5Encrypt32(pass));
             if (user != null)
             {
 
-                TokenModelJWT tokenModel = new TokenModelJWT();
-                tokenModel.Uid = 1;
-                tokenModel.Role = user;
+                TokenModelJwt tokenModel = new TokenModelJwt { Uid = 1, Role = user };
 
-                jwtStr = JwtHelper.IssueJWT(tokenModel);
+                jwtStr = JwtHelper.IssueJwt(tokenModel);
                 suc = true;
             }
             else
@@ -83,10 +83,15 @@ namespace Blog.Core.Controllers
         }
 
 
-
+        /// <summary>
+        /// 获取JWT的方法2：给Nuxt提供
+        /// </summary>
+        /// <param name="name"></param>
+        /// <param name="pass"></param>
+        /// <returns></returns>
         [HttpGet]
         [Route("GetTokenNuxt")]
-        public async Task<object> GetJWTStrForNuxt(string name, string pass)
+        public object GetJwtStrForNuxt(string name, string pass)
         {
             string jwtStr = string.Empty;
             bool suc = false;
@@ -94,11 +99,13 @@ namespace Blog.Core.Controllers
             //这里直接写死了
             if (name == "admins" && pass == "admins")
             {
-                TokenModelJWT tokenModel = new TokenModelJWT();
-                tokenModel.Uid = 1;
-                tokenModel.Role = "Admin";
+                TokenModelJwt tokenModel = new TokenModelJwt
+                {
+                    Uid = 1,
+                    Role = "Admin"
+                };
 
-                jwtStr = JwtHelper.IssueJWT(tokenModel);
+                jwtStr = JwtHelper.IssueJwt(tokenModel);
                 suc = true;
             }
             else
@@ -121,17 +128,16 @@ namespace Blog.Core.Controllers
 
 
         /// <summary>
-        /// 获取JWT的方法 3.0
+        /// 获取JWT的方法3：整个系统主要方法
         /// </summary>
         /// <param name="name"></param>
         /// <param name="pass"></param>
         /// <returns></returns>
         [HttpGet]
         [Route("JWTToken3.0")]
-        public async Task<object> GetJWTToken3(string name = "", string pass = "")
+        public async Task<object> GetJwtToken3(string name = "", string pass = "")
         {
             string jwtStr = string.Empty;
-            bool suc = false;
 
             if (string.IsNullOrEmpty(name) || string.IsNullOrEmpty(pass))
             {
@@ -142,10 +148,12 @@ namespace Blog.Core.Controllers
                 });
             }
 
-            var user = await sysUserInfoServices.Query(d => d.uLoginName == name && d.uLoginPWD == pass);
+            pass = MD5Helper.MD5Encrypt32(pass);
+
+            var user = await _sysUserInfoServices.Query(d => d.uLoginName == name && d.uLoginPWD == pass);
             if (user.Count > 0)
             {
-                var userRoles = await sysUserInfoServices.GetUserRoleNameStr(name, pass);
+                var userRoles = await _sysUserInfoServices.GetUserRoleNameStr(name, pass);
                 //如果是基于用户的授权策略，这里要添加用户;如果是基于角色的授权策略，这里要添加角色
                 var claims = new List<Claim> {
                     new Claim(ClaimTypes.Name, name),
@@ -173,10 +181,59 @@ namespace Blog.Core.Controllers
 
         }
 
+        /// <summary>
+        /// 请求刷新Token（以旧换新）
+        /// </summary>
+        /// <param name="token"></param>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("RefreshToken")]
+        public async Task<object> RefreshToken(string token = "")
+        {
+            string jwtStr = string.Empty;
+
+            if (string.IsNullOrEmpty(token))
+            {
+                return new JsonResult(new
+                {
+                    Status = false,
+                    message = "token无效，请重新登录！"
+                });
+            }
+            var tokenModel = JwtHelper.SerializeJwt(token);
+            if (tokenModel != null && tokenModel.Uid > 0)
+            {
+                var user = await _sysUserInfoServices.QueryById(tokenModel.Uid);
+                if (user != null)
+                {
+                    var userRoles = await _sysUserInfoServices.GetUserRoleNameStr(user.uLoginName, user.uLoginPWD);
+                    //如果是基于用户的授权策略，这里要添加用户;如果是基于角色的授权策略，这里要添加角色
+                    var claims = new List<Claim> {
+                    new Claim(ClaimTypes.Name, user.uLoginName),
+                    new Claim(JwtRegisteredClaimNames.Jti, tokenModel.Uid.ObjToString()),
+                    new Claim(ClaimTypes.Expiration, DateTime.Now.AddSeconds(_requirement.Expiration.TotalSeconds).ToString()) };
+                    claims.AddRange(userRoles.Split(',').Select(s => new Claim(ClaimTypes.Role, s)));
+
+                    //用户标识
+                    var identity = new ClaimsIdentity(JwtBearerDefaults.AuthenticationScheme);
+                    identity.AddClaims(claims);
+
+                    var refreshToken = JwtToken.BuildJwtToken(claims.ToArray(), _requirement);
+                    return new JsonResult(refreshToken);
+                }
+            }
+
+            return new JsonResult(new
+            {
+                success = false,
+                message = "认证失败"
+            });
+        }
 
         /// <summary>
-        /// 
+        /// 获取JWT的方法4：给 JSONP 测试
         /// </summary>
+        /// <param name="callBack"></param>
         /// <param name="id"></param>
         /// <param name="sub"></param>
         /// <param name="expiresSliding"></param>
@@ -186,15 +243,30 @@ namespace Blog.Core.Controllers
         [Route("jsonp")]
         public void Getjsonp(string callBack, long id = 1, string sub = "Admin", int expiresSliding = 30, int expiresAbsoulute = 30)
         {
-            TokenModelJWT tokenModel = new TokenModelJWT();
-            tokenModel.Uid = id;
-            tokenModel.Role = sub;
+            TokenModelJwt tokenModel = new TokenModelJwt
+            {
+                Uid = id,
+                Role = sub
+            };
 
-            string jwtStr = JwtHelper.IssueJWT(tokenModel);
+            string jwtStr = JwtHelper.IssueJwt(tokenModel);
 
             string response = string.Format("\"value\":\"{0}\"", jwtStr);
             string call = callBack + "({" + response + "})";
             Response.WriteAsync(call);
+        }
+
+
+        /// <summary>
+        /// 测试 MD5 加密字符串
+        /// </summary>
+        /// <param name="password"></param>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("Md5Password")]
+        public string Md5Password(string password = "")
+        {
+            return MD5Helper.MD5Encrypt32(password);
         }
     }
 }
