@@ -1,68 +1,60 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
+using Blog.Core.Common.HttpContextUser;
 using Blog.Core.IServices;
 using Blog.Core.Model;
 using Blog.Core.Model.Models;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Blog.Core.Controllers
 {
+    /// <summary>
+    /// 接口管理
+    /// </summary>
     [Route("api/[controller]/[action]")]
     [ApiController]
-    [Authorize("Permission")]
+    [Authorize(Permissions.Name)]
     public class ModuleController : ControllerBase
     {
-        IModuleServices _ModuleServices;
+        readonly IModuleServices _moduleServices;
+        readonly IUser _user;
 
-        /// <summary>
-        /// 构造函数
-        /// </summary>
-        /// <param name="ModuleServices"></param>
-        public ModuleController(IModuleServices ModuleServices )
+       
+        public ModuleController(IModuleServices moduleServices, IUser user)
         {
-            _ModuleServices = ModuleServices;
+            _moduleServices = moduleServices;
+            _user = user;
         }
 
+        /// <summary>
+        /// 获取全部接口api
+        /// </summary>
+        /// <param name="page"></param>
+        /// <param name="key"></param>
+        /// <returns></returns>
         // GET: api/User
         [HttpGet]
         public async Task<MessageModel<PageModel<Module>>> Get(int page = 1, string key = "")
         {
-            var data = new MessageModel<PageModel<Module>>();
-            int intTotalCount = 50;
-            int TotalCount = 0;
-            int PageCount = 1;
-            List<Module> Modules = new List<Module>();
-
-            Modules = await _ModuleServices.Query(a => a.IsDeleted != true );
-
-            if (!string.IsNullOrEmpty(key))
+            if (string.IsNullOrEmpty(key) || string.IsNullOrWhiteSpace(key))
             {
-                Modules = Modules.Where(t => (t.Name != null && t.Name.Contains(key))).ToList();
+                key = "";
             }
+            int intPageSize = 50;
 
+            Expression<Func<Module, bool>> whereExpression = a => a.IsDeleted != true && (a.Name != null && a.Name.Contains(key));
 
-            //筛选后的数据总数
-            TotalCount = Modules.Count;
-            //筛选后的总页数
-            PageCount = (Math.Ceiling(TotalCount.ObjToDecimal() / intTotalCount.ObjToDecimal())).ObjToInt();
-
-            Modules = Modules.OrderByDescending(d => d.Id).Skip((page - 1) * intTotalCount).Take(intTotalCount).ToList();
+            var data = await _moduleServices.QueryPage(whereExpression, page, intPageSize, " Id desc ");
 
             return new MessageModel<PageModel<Module>>()
             {
                 msg = "获取成功",
-                success = TotalCount >= 0,
-                response = new PageModel<Module>()
-                {
-                    page = page,
-                    pageCount = PageCount,
-                    dataCount = TotalCount,
-                    data = Modules,
-                }
+                success = data.dataCount >= 0,
+                response = data
             };
 
         }
@@ -74,13 +66,21 @@ namespace Blog.Core.Controllers
             return "value";
         }
 
+        /// <summary>
+        /// 添加一条接口信息
+        /// </summary>
+        /// <param name="module"></param>
+        /// <returns></returns>
         // POST: api/User
         [HttpPost]
-        public async Task<MessageModel<string>> Post([FromBody] Module Module)
+        public async Task<MessageModel<string>> Post([FromBody] Module module)
         {
             var data = new MessageModel<string>();
 
-            var id = (await _ModuleServices.Add(Module));
+            module.CreateId = _user.ID;
+            module.CreateBy = _user.Name;
+
+            var id = (await _moduleServices.Add(module));
             data.success = id > 0;
             if (data.success)
             {
@@ -91,24 +91,34 @@ namespace Blog.Core.Controllers
             return data;
         }
 
+        /// <summary>
+        /// 更新接口信息
+        /// </summary>
+        /// <param name="module"></param>
+        /// <returns></returns>
         // PUT: api/User/5
         [HttpPut]
-        public async Task<MessageModel<string>> Put([FromBody] Module Module)
+        public async Task<MessageModel<string>> Put([FromBody] Module module)
         {
             var data = new MessageModel<string>();
-            if (Module != null && Module.Id > 0)
+            if (module != null && module.Id > 0)
             {
-                data.success = await _ModuleServices.Update(Module);
+                data.success = await _moduleServices.Update(module);
                 if (data.success)
                 {
                     data.msg = "更新成功";
-                    data.response = Module?.Id.ObjToString();
+                    data.response = module?.Id.ObjToString();
                 }
             }
 
             return data;
         }
 
+        /// <summary>
+        /// 删除一条接口
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         // DELETE: api/ApiWithActions/5
         [HttpDelete]
         public async Task<MessageModel<string>> Delete(int id)
@@ -116,9 +126,9 @@ namespace Blog.Core.Controllers
             var data = new MessageModel<string>();
             if (id > 0)
             {
-                var userDetail = await _ModuleServices.QueryByID(id);
+                var userDetail = await _moduleServices.QueryById(id);
                 userDetail.IsDeleted = true;
-                data.success = await _ModuleServices.Update(userDetail);
+                data.success = await _moduleServices.Update(userDetail);
                 if (data.success)
                 {
                     data.msg = "删除成功";
