@@ -2,7 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Blog.Core.AuthHelper.OverWrite;
 using Blog.Core.Common.Helper;
+using Blog.Core.Common.HttpContextUser;
 using Blog.Core.IServices;
 using Blog.Core.Model;
 using Blog.Core.Model.Models;
@@ -12,79 +14,105 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Blog.Core.Controllers
 {
+    /// <summary>
+    /// 菜单管理
+    /// </summary>
     [Route("api/[controller]/[action]")]
     [ApiController]
-    [Authorize("Permission")]
+    [Authorize(Permissions.Name)]
     public class PermissionController : ControllerBase
     {
-        IPermissionServices _PermissionServices;
-        IModuleServices _ModuleServices;
-        IRoleModulePermissionServices _roleModulePermissionServices;
-        IUserRoleServices _userRoleServices;
+        readonly IPermissionServices _permissionServices;
+        readonly IModuleServices _moduleServices;
+        readonly IRoleModulePermissionServices _roleModulePermissionServices;
+        readonly IUserRoleServices _userRoleServices;
+        readonly IHttpContextAccessor _httpContext;
+        readonly IUser _user;
 
         /// <summary>
         /// 构造函数
         /// </summary>
-        /// <param name="PermissionServices"></param>
-        /// <param name="ModuleServices"></param>
+        /// <param name="permissionServices"></param>
+        /// <param name="moduleServices"></param>
         /// <param name="roleModulePermissionServices"></param>
         /// <param name="userRoleServices"></param>
-        public PermissionController(IPermissionServices PermissionServices, IModuleServices ModuleServices, IRoleModulePermissionServices roleModulePermissionServices, IUserRoleServices userRoleServices)
+        /// <param name="httpContext"></param>
+        /// <param name="user"></param>
+        public PermissionController(IPermissionServices permissionServices, IModuleServices moduleServices, IRoleModulePermissionServices roleModulePermissionServices, IUserRoleServices userRoleServices, IHttpContextAccessor httpContext, IUser user)
         {
-            _PermissionServices = PermissionServices;
-            _ModuleServices = ModuleServices;
+            _permissionServices = permissionServices;
+            _moduleServices = moduleServices;
             _roleModulePermissionServices = roleModulePermissionServices;
             _userRoleServices = userRoleServices;
+            _httpContext = httpContext;
+            _user = user;
 
         }
 
+        /// <summary>
+        /// 获取菜单
+        /// </summary>
+        /// <param name="page"></param>
+        /// <param name="key"></param>
+        /// <returns></returns>
         // GET: api/User
         [HttpGet]
         public async Task<MessageModel<PageModel<Permission>>> Get(int page = 1, string key = "")
         {
-            var data = new MessageModel<PageModel<Permission>>();
-            int intTotalCount = 50;
-            int TotalCount = 0;
-            int PageCount = 1;
-            List<Permission> Permissions = new List<Permission>();
-
-            Permissions = await _PermissionServices.Query(a => a.IsDeleted != true);
-
-            if (!string.IsNullOrEmpty(key))
+            PageModel<Permission> permissions = new PageModel<Permission>();
+            int intPageSize = 50;
+            if (string.IsNullOrEmpty(key) || string.IsNullOrWhiteSpace(key))
             {
-                Permissions = Permissions.Where(t => (t.Name != null && t.Name.Contains(key))).ToList();
+                key = "";
             }
 
+            #region 舍弃
+            //var permissions = await _permissionServices.Query(a => a.IsDeleted != true);
+            //if (!string.IsNullOrEmpty(key))
+            //{
+            //    permissions = permissions.Where(t => (t.Name != null && t.Name.Contains(key))).ToList();
+            //}
+            ////筛选后的数据总数
+            //totalCount = permissions.Count;
+            ////筛选后的总页数
+            //pageCount = (Math.Ceiling(totalCount.ObjToDecimal() / intTotalCount.ObjToDecimal())).ObjToInt();
+            //permissions = permissions.OrderByDescending(d => d.Id).Skip((page - 1) * intTotalCount).Take(intTotalCount).ToList(); 
+            #endregion
 
-            //筛选后的数据总数
-            TotalCount = Permissions.Count;
-            //筛选后的总页数
-            PageCount = (Math.Ceiling(TotalCount.ObjToDecimal() / intTotalCount.ObjToDecimal())).ObjToInt();
 
-            Permissions = Permissions.OrderByDescending(d => d.Id).Skip((page - 1) * intTotalCount).Take(intTotalCount).ToList();
-            var apis = await _ModuleServices.Query(d => d.IsDeleted == false);
 
-            foreach (var item in Permissions)
+            permissions = await _permissionServices.QueryPage(a => a.IsDeleted != true && (a.Name != null && a.Name.Contains(key)), page, intPageSize, " Id desc ");
+
+
+            #region 单独处理
+
+            var apis = await _moduleServices.Query(d => d.IsDeleted == false);
+            var permissionsView = permissions.data;
+
+            var permissionAll = await _permissionServices.Query(d => d.IsDeleted != true);
+            foreach (var item in permissionsView)
             {
-                List<int> pidarr = new List<int>();
-                pidarr.Add(item.Pid);
+                List<int> pidarr = new List<int>
+                {
+                    item.Pid
+                };
                 if (item.Pid > 0)
                 {
                     pidarr.Add(0);
                 }
-                var parent = Permissions.Where(d => d.Id == item.Pid).FirstOrDefault();
+                var parent = permissionAll.FirstOrDefault(d => d.Id == item.Pid);
 
                 while (parent != null)
                 {
                     pidarr.Add(parent.Id);
-                    parent = Permissions.Where(d => d.Id == parent.Pid).FirstOrDefault();
+                    parent = permissionAll.FirstOrDefault(d => d.Id == parent.Pid);
                 }
 
 
                 item.PidArr = pidarr.OrderBy(d => d).Distinct().ToList();
                 foreach (var pid in item.PidArr)
                 {
-                    var per = Permissions.Where(d => d.Id == pid).FirstOrDefault();
+                    var per = permissionAll.FirstOrDefault(d => d.Id == pid);
                     item.PnameArr.Add((per != null ? per.Name : "根节点") + "/");
                     //var par = Permissions.Where(d => d.Pid == item.Id ).ToList();
                     //item.PCodeArr.Add((per != null ? $"/{per.Code}/{item.Code}" : ""));
@@ -94,20 +122,19 @@ namespace Blog.Core.Controllers
                     //}
                 }
 
-                item.MName = apis.Where(d => d.Id == item.Mid).FirstOrDefault()?.LinkUrl;
+                item.MName = apis.FirstOrDefault(d => d.Id == item.Mid)?.LinkUrl;
             }
+
+            permissions.data = permissionsView;
+
+            #endregion
+
 
             return new MessageModel<PageModel<Permission>>()
             {
                 msg = "获取成功",
-                success = TotalCount >= 0,
-                response = new PageModel<Permission>()
-                {
-                    page = page,
-                    pageCount = PageCount,
-                    dataCount = TotalCount,
-                    data = Permissions,
-                }
+                success = permissions.dataCount >= 0,
+                response = permissions
             };
 
         }
@@ -119,13 +146,21 @@ namespace Blog.Core.Controllers
             return "value";
         }
 
+        /// <summary>
+        /// 添加一个菜单
+        /// </summary>
+        /// <param name="permission"></param>
+        /// <returns></returns>
         // POST: api/User
         [HttpPost]
-        public async Task<MessageModel<string>> Post([FromBody] Permission Permission)
+        public async Task<MessageModel<string>> Post([FromBody] Permission permission)
         {
             var data = new MessageModel<string>();
 
-            var id = (await _PermissionServices.Add(Permission));
+            permission.CreateId = _user.ID;
+            permission.CreateBy = _user.Name;
+
+            var id = (await _permissionServices.Add(permission));
             data.success = id > 0;
             if (data.success)
             {
@@ -136,7 +171,11 @@ namespace Blog.Core.Controllers
             return data;
         }
 
-
+        /// <summary>
+        /// 保存菜单权限分配
+        /// </summary>
+        /// <param name="assignView"></param>
+        /// <returns></returns>
         [HttpPost]
         public async Task<MessageModel<string>> Assign([FromBody] AssignView assignView)
         {
@@ -157,9 +196,9 @@ namespace Blog.Core.Controllers
                     foreach (var item in assignView.pids)
                     {
                         var rmpitem = roleModulePermissions.Where(d => d.PermissionId == item);
-                        if (rmpitem.Count() == 0)
+                        if (!rmpitem.Any())
                         {
-                            var moduleid = (await _PermissionServices.Query(p => p.Id == item)).FirstOrDefault()?.Mid;
+                            var moduleid = (await _permissionServices.Query(p => p.Id == item)).FirstOrDefault()?.Mid;
                             RoleModulePermission roleModulePermission = new RoleModulePermission()
                             {
                                 IsDeleted = false,
@@ -167,6 +206,10 @@ namespace Blog.Core.Controllers
                                 ModuleId = moduleid.ObjToInt(),
                                 PermissionId = item,
                             };
+
+
+                            roleModulePermission.CreateId = _user.ID;
+                            roleModulePermission.CreateBy = _user.Name;
 
                             data.success |= (await _roleModulePermissionServices.Add(roleModulePermission)) > 0;
 
@@ -189,12 +232,19 @@ namespace Blog.Core.Controllers
             return data;
         }
 
+
+        /// <summary>
+        /// 获取菜单树
+        /// </summary>
+        /// <param name="pid"></param>
+        /// <param name="needbtn"></param>
+        /// <returns></returns>
         [HttpGet]
         public async Task<MessageModel<PermissionTree>> GetPermissionTree(int pid = 0, bool needbtn = false)
         {
             var data = new MessageModel<PermissionTree>();
 
-            var permissions = await _PermissionServices.Query(d => d.IsDeleted == false);
+            var permissions = await _permissionServices.Query(d => d.IsDeleted == false);
             var permissionTrees = (from child in permissions
                                    where child.IsDeleted == false
                                    orderby child.Id
@@ -206,10 +256,12 @@ namespace Blog.Core.Controllers
                                        isbtn = child.IsButton,
                                        order = child.OrderSort,
                                    }).ToList();
-            PermissionTree rootRoot = new PermissionTree();
-            rootRoot.value = 0;
-            rootRoot.Pid = 0;
-            rootRoot.label = "根节点";
+            PermissionTree rootRoot = new PermissionTree
+            {
+                value = 0,
+                Pid = 0,
+                label = "根节点"
+            };
 
             permissionTrees = permissionTrees.OrderBy(d => d.order).ToList();
 
@@ -226,23 +278,37 @@ namespace Blog.Core.Controllers
             return data;
         }
 
-
+        /// <summary>
+        /// 获取路由树
+        /// </summary>
+        /// <param name="uid"></param>
+        /// <returns></returns>
         [HttpGet]
-        [AllowAnonymous]
         public async Task<MessageModel<NavigationBar>> GetNavigationBar(int uid)
         {
+
             var data = new MessageModel<NavigationBar>();
 
-            if (uid > 0)
+            // 三种方式获取 uid
+            var uidInHttpcontext1 = (from item in _httpContext.HttpContext.User.Claims
+                                     where item.Type == "jti"
+                                     select item.Value).FirstOrDefault().ObjToInt();
+
+            var uidInHttpcontext = (JwtHelper.SerializeJwt(_httpContext.HttpContext.Request.Headers["Authorization"].ObjToString().Replace("Bearer ", "")))?.Uid;
+
+            var uName = _user.Name;
+
+            if (uid > 0 && uid == uidInHttpcontext)
             {
                 var roleId = ((await _userRoleServices.Query(d => d.IsDeleted == false && d.UserId == uid)).FirstOrDefault()?.RoleId).ObjToInt();
                 if (roleId > 0)
                 {
                     var pids = (await _roleModulePermissionServices.Query(d => d.IsDeleted == false && d.RoleId == roleId)).Select(d => d.PermissionId.ObjToInt()).Distinct();
 
-                    if (pids.Count() > 0)
+                    if (pids.Any())
                     {
-                        var rolePermissionMoudles = (await _PermissionServices.Query(d => pids.Contains(d.Id) && d.IsButton == false)).OrderBy(c => c.OrderSort);
+                        //var rolePermissionMoudles = (await _permissionServices.Query(d => pids.Contains(d.Id) && d.IsButton == false)).OrderBy(c => c.OrderSort);
+                        var rolePermissionMoudles = (await _permissionServices.Query(d => pids.Contains(d.Id))).OrderBy(c => c.OrderSort);
                         var permissionTrees = (from child in rolePermissionMoudles
                                                where child.IsDeleted == false
                                                orderby child.Id
@@ -254,10 +320,14 @@ namespace Blog.Core.Controllers
                                                    order = child.OrderSort,
                                                    path = child.Code,
                                                    iconCls = child.Icon,
+                                                   Func=child.Func,
+                                                   IsHide = child.IsHide.ObjToBool(),
+                                                   IsButton=child.IsButton.ObjToBool(),
                                                    meta = new NavigationBarMeta
                                                    {
                                                        requireAuth = true,
                                                        title = child.Name,
+                                                       NoTabPage = child.IsHide.ObjToBool()
                                                    }
                                                }).ToList();
 
@@ -290,7 +360,11 @@ namespace Blog.Core.Controllers
             return data;
         }
 
-
+        /// <summary>
+        /// 通过角色获取菜单【无权限】
+        /// </summary>
+        /// <param name="rid"></param>
+        /// <returns></returns>
         [HttpGet]
         [AllowAnonymous]
         public async Task<MessageModel<AssignShow>> GetPermissionIdByRoleId(int rid = 0)
@@ -302,12 +376,12 @@ namespace Blog.Core.Controllers
                                    orderby child.Id
                                    select child.PermissionId.ObjToInt()).ToList();
 
-            var permissions = await _PermissionServices.Query(d => d.IsDeleted == false);
+            var permissions = await _permissionServices.Query(d => d.IsDeleted == false);
             List<string> assignbtns = new List<string>();
 
             foreach (var item in permissionTrees)
             {
-                var pername = permissions.Where(d => d.IsButton && d.Id == item).FirstOrDefault()?.Name;
+                var pername = permissions.FirstOrDefault(d => d.IsButton && d.Id == item)?.Name;
                 if (!string.IsNullOrEmpty(pername))
                 {
                     assignbtns.Add(pername + "_" + item);
@@ -328,25 +402,34 @@ namespace Blog.Core.Controllers
             return data;
         }
 
-
+        /// <summary>
+        /// 更新菜单
+        /// </summary>
+        /// <param name="permission"></param>
+        /// <returns></returns>
         // PUT: api/User/5
         [HttpPut]
-        public async Task<MessageModel<string>> Put([FromBody] Permission Permission)
+        public async Task<MessageModel<string>> Put([FromBody] Permission permission)
         {
             var data = new MessageModel<string>();
-            if (Permission != null && Permission.Id > 0)
+            if (permission != null && permission.Id > 0)
             {
-                data.success = await _PermissionServices.Update(Permission);
+                data.success = await _permissionServices.Update(permission);
                 if (data.success)
                 {
                     data.msg = "更新成功";
-                    data.response = Permission?.Id.ObjToString();
+                    data.response = permission?.Id.ObjToString();
                 }
             }
 
             return data;
         }
 
+        /// <summary>
+        /// 删除菜单
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         // DELETE: api/ApiWithActions/5
         [HttpDelete]
         public async Task<MessageModel<string>> Delete(int id)
@@ -354,9 +437,9 @@ namespace Blog.Core.Controllers
             var data = new MessageModel<string>();
             if (id > 0)
             {
-                var userDetail = await _PermissionServices.QueryByID(id);
+                var userDetail = await _permissionServices.QueryById(id);
                 userDetail.IsDeleted = true;
-                data.success = await _PermissionServices.Update(userDetail);
+                data.success = await _permissionServices.Update(userDetail);
                 if (data.success)
                 {
                     data.msg = "删除成功";
