@@ -2,24 +2,29 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Blog.Core.AuthHelper;
+using Blog.Core.AuthHelper.OverWrite;
+using Blog.Core.Common.Helper;
+using Blog.Core.Common.HttpContextUser;
 using Blog.Core.IServices;
 using Blog.Core.Model;
 using Blog.Core.Model.Models;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Blog.Core.Controllers
 {
+    /// <summary>
+    /// 用户管理
+    /// </summary>
     [Route("api/[controller]/[action]")]
     [ApiController]
-    [Authorize("Permission")]
+    [Authorize(Permissions.Name)]
     public class UserController : ControllerBase
     {
-        IsysUserInfoServices _sysUserInfoServices;
-        IUserRoleServices _userRoleServices;
-        IRoleServices _roleServices;
+        readonly ISysUserInfoServices _sysUserInfoServices;
+        readonly IUserRoleServices _userRoleServices;
+        readonly IRoleServices _roleServices;
+        private readonly IUser _user;
 
         /// <summary>
         /// 构造函数
@@ -27,61 +32,56 @@ namespace Blog.Core.Controllers
         /// <param name="sysUserInfoServices"></param>
         /// <param name="userRoleServices"></param>
         /// <param name="roleServices"></param>
-        public UserController(IsysUserInfoServices sysUserInfoServices, IUserRoleServices userRoleServices, IRoleServices roleServices)
+        /// <param name="user"></param>
+        public UserController(ISysUserInfoServices sysUserInfoServices, IUserRoleServices userRoleServices, IRoleServices roleServices, IUser user)
         {
             _sysUserInfoServices = sysUserInfoServices;
             _userRoleServices = userRoleServices;
             _roleServices = roleServices;
+            _user = user;
         }
 
+        /// <summary>
+        /// 获取全部用户
+        /// </summary>
+        /// <param name="page"></param>
+        /// <param name="key"></param>
+        /// <returns></returns>
         // GET: api/User
         [HttpGet]
+        [ResponseCache(Duration = 60)]
         public async Task<MessageModel<PageModel<sysUserInfo>>> Get(int page = 1, string key = "")
         {
-            var data = new MessageModel<PageModel<sysUserInfo>>();
-            int intTotalCount = 50;
-            int TotalCount = 0;
-            int PageCount = 1;
-            List<sysUserInfo> sysUserInfos = new List<sysUserInfo>();
-
-            sysUserInfos = await _sysUserInfoServices.Query(a => a.tdIsDelete != true && a.uStatus >= 0);
-
-            if (!string.IsNullOrEmpty(key))
+            if (string.IsNullOrEmpty(key) || string.IsNullOrWhiteSpace(key))
             {
-                sysUserInfos = sysUserInfos.Where(t => (t.uLoginName != null && t.uLoginName.Contains(key)) || (t.uRealName != null && t.uRealName.Contains(key))).ToList();
+                key = "";
             }
+            int intPageSize = 50;
 
 
-            //筛选后的数据总数
-            TotalCount = sysUserInfos.Count;
-            //筛选后的总页数
-            PageCount = (Math.Ceiling(TotalCount.ObjToDecimal() / intTotalCount.ObjToDecimal())).ObjToInt();
-
-            sysUserInfos = sysUserInfos.OrderByDescending(d => d.uID).Skip((page - 1) * intTotalCount).Take(intTotalCount).ToList();
+            var data = await _sysUserInfoServices.QueryPage(a => a.tdIsDelete != true && a.uStatus >= 0 && ((a.uLoginName != null && a.uLoginName.Contains(key)) || (a.uRealName != null && a.uRealName.Contains(key))), page, intPageSize, " uID desc ");
 
 
+            #region MyRegion
             var allUserRoles = await _userRoleServices.Query(d => d.IsDeleted == false);
             var allRoles = await _roleServices.Query(d => d.IsDeleted == false);
+
+            var sysUserInfos = data.data;
             foreach (var item in sysUserInfos)
             {
-                if (item != null)
-                {
-                    item.RID = (allUserRoles.Where(d => d.UserId == item.uID).FirstOrDefault()?.RoleId).ObjToInt();
-                    item.RoleName = allRoles.Where(d=>d.Id==item.RID).FirstOrDefault()?.Name;
-                }
+                item.RID = (allUserRoles.FirstOrDefault(d => d.UserId == item.uID)?.RoleId).ObjToInt();
+                item.RoleName = allRoles.FirstOrDefault(d => d.Id == item.RID)?.Name;
             }
+
+            data.data = sysUserInfos; 
+            #endregion
+
 
             return new MessageModel<PageModel<sysUserInfo>>()
             {
                 msg = "获取成功",
-                success = TotalCount >= 0,
-                response = new PageModel<sysUserInfo>()
-                {
-                    page = page,
-                    pageCount = PageCount,
-                    dataCount = TotalCount,
-                    data = sysUserInfos,
-                }
+                success = data.dataCount >= 0,
+                response = data
             };
 
         }
@@ -96,6 +96,7 @@ namespace Blog.Core.Controllers
         // GET: api/User/5
         /// <summary>
         /// 获取用户详情根据token
+        /// 【无权限】
         /// </summary>
         /// <param name="token">令牌</param>
         /// <returns></returns>
@@ -106,10 +107,10 @@ namespace Blog.Core.Controllers
             var data = new MessageModel<sysUserInfo>();
             if (!string.IsNullOrEmpty(token))
             {
-                var tokenModel = JwtHelper.SerializeJWT(token);
+                var tokenModel = JwtHelper.SerializeJwt(token);
                 if (tokenModel != null && tokenModel.Uid > 0)
                 {
-                    var userinfo = await _sysUserInfoServices.QueryByID(tokenModel.Uid);
+                    var userinfo = await _sysUserInfoServices.QueryById(tokenModel.Uid);
                     if (userinfo != null)
                     {
                         data.response = userinfo;
@@ -122,11 +123,19 @@ namespace Blog.Core.Controllers
             return data;
         }
 
+        /// <summary>
+        /// 添加一个用户
+        /// </summary>
+        /// <param name="sysUserInfo"></param>
+        /// <returns></returns>
         // POST: api/User
         [HttpPost]
         public async Task<MessageModel<string>> Post([FromBody] sysUserInfo sysUserInfo)
         {
             var data = new MessageModel<string>();
+
+            sysUserInfo.uLoginPWD= MD5Helper.MD5Encrypt32(sysUserInfo.uLoginPWD);
+            sysUserInfo.uRemark = _user.Name;
 
             var id = await _sysUserInfoServices.Add(sysUserInfo);
             data.success = id > 0;
@@ -139,10 +148,17 @@ namespace Blog.Core.Controllers
             return data;
         }
 
+        /// <summary>
+        /// 更新用户与角色
+        /// </summary>
+        /// <param name="sysUserInfo"></param>
+        /// <returns></returns>
         // PUT: api/User/5
         [HttpPut]
         public async Task<MessageModel<string>> Put([FromBody] sysUserInfo sysUserInfo)
         {
+            // 这里也要做后期处理，会有用户个人中心的业务
+
             var data = new MessageModel<string>();
             if (sysUserInfo != null && sysUserInfo.uID > 0)
             {
@@ -166,6 +182,11 @@ namespace Blog.Core.Controllers
             return data;
         }
 
+        /// <summary>
+        /// 删除用户
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         // DELETE: api/ApiWithActions/5
         [HttpDelete]
         public async Task<MessageModel<string>> Delete(int id)
@@ -173,7 +194,7 @@ namespace Blog.Core.Controllers
             var data = new MessageModel<string>();
             if (id > 0)
             {
-                var userDetail = await _sysUserInfoServices.QueryByID(id);
+                var userDetail = await _sysUserInfoServices.QueryById(id);
                 userDetail.tdIsDelete = true;
                 data.success = await _sysUserInfoServices.Update(userDetail);
                 if (data.success)
