@@ -10,10 +10,10 @@ using static Blog.Core.SwaggerHelper.CustomApiVersion;
 
 namespace Blog.Core.Controllers.v2
 {
-    [CustomRoute(ApiVersions.v2)]
+    [CustomRoute(ApiVersions.V2)]
     //[Route("api/[controller]")]
     [ApiController]
-    [Authorize("Permission")]
+    [Authorize(Permissions.Name)]
     public class ApbController : ControllerBase
     {
 
